重构产品页静态页面文件结构 将所有props配置抽到顶层 而不是分组件进行对应语言配置
