+++ apps/gemebuild/src/i18n-pages/pdp/components/UserStory.tsx
@@ -1,0 +1,12 @@
+import { PdpUserStoryTeamSections } from 'ui'
+import type { IPdpUserStoryTeamSectionsProps } from 'ui'
+
+function UserStory({
+  pdpUserStoryTeamSectionsProps,
+}: {
+  pdpUserStoryTeamSectionsProps: IPdpUserStoryTeamSectionsProps
+}) {
+  return <PdpUserStoryTeamSections {...pdpUserStoryTeamSectionsProps} />
+}
+
+export default UserStory
