+++ apps/gemebuild/src/i18n-pages/pdp/index.ts
@@ -1,3 +1,3 @@
-export { default as PdpIntroEn } from './pdp-intro-en'
-export { default as PdpIntroDe } from './pdp-intro-de'
-export { default as PdpIntroFr } from './pdp-intro-fr'
+export * from './en'
+export * from './de'
+export * from './fr'
